A-Packages: Added duke, commands and tasks packages
