+++ src/main/java/duke/Duke.java
@@ -1,3 +1,7 @@
+package duke;
+
+import duke.commands.Command;
+
 import java.util.Scanner;
 
 public class Duke {
