+++ src/main/java/duke/Parser.java
@@ -1,3 +1,13 @@
+package duke;
+
+import duke.commands.Command;
+import duke.commands.ByeCommand;
+import duke.commands.ListCommand;
+import duke.commands.NumericCommand;
+import duke.commands.TodoCommand;
+import duke.commands.DeadlineCommand;
+import duke.commands.EventCommand;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Arrays;
