+++ src/main/java/duke/commands/ByeCommand.java
@@ -1,5 +1,8 @@
-import java.io.FileWriter;
-import java.io.IOException;
+package duke.commands;
+
+import duke.TaskList;
+import duke.Storage;
+import duke.DukeException;
 
 public class ByeCommand implements Command {
