+++ src/main/java/duke/commands/Command.java
@@ -1,3 +1,8 @@
+package duke.commands;
+import duke.TaskList;
+import duke.Storage;
+import duke.DukeException;
+
 public interface Command {
     public void execute(TaskList tasks, Storage storage) throws DukeException;
 }
