+++ src/main/java/duke/tasks/TaskType.java
@@ -1,3 +1,4 @@
+package duke.tasks;
 public enum TaskType {
     T,
     D,
