+++ text-ui-test/EXPECTED.TXT
@@ -1,6 +1,6 @@
-Hello from Duke
+Hello from duke
 What can I do for you?
 Got it. I've added this task:
 [T][ ]  borrow book
-Now you have 1 tasks in the list
+Now you have 1 duke.tasks in the list
 Bye. Hope to see you again soon!
