+++ text-ui-test/complete test/expected.txt
@@ -1,15 +1,15 @@
-Hello from Duke
+Hello from duke
 What can I do for you?
 Got it. I've added this task:
 [T][ ]  borrow book
-Now you have 1 tasks in the list
+Now you have 1 duke.tasks in the list
 1. [T][ ]  borrow book
 Got it. I've added this task:
 [D][ ]  return book (by: Sunday)
-Now you have 2 tasks in the list
+Now you have 2 duke.tasks in the list
 Got it. I've added this task:
 [E][ ]  project meeting (at: Mon 2-4pm)
-Now you have 3 tasks in the list
+Now you have 3 duke.tasks in the list
 1. [T][ ]  borrow book
 2. [D][ ]  return book (by: Sunday)
 3. [E][ ]  project meeting (at: Mon 2-4pm)
